Major refactoring and switch to function vectors for various commands
