+++ src/drive.c
@@ -1,0 +1,3 @@
+#ifndef ARDUINO
+   #include "drive.cpp"
+#endif
