+++ src/hexbus.h
@@ -24,6 +24,8 @@
 extern "C"{
 #endif
 
+#include "integer.h"
+
 typedef enum _hexcmdtype_t {
                HEXCMD_OPEN = 0,
                HEXCMD_CLOSE,
@@ -114,7 +116,7 @@ typedef enum _openmode_t {
             } openmode_t ;
 
 uint8_t hex_is_bav(void);
-void hex_release_data( void );
+//void hex_release_data( void );
 void hex_release_bus(void);
 uint8_t hex_capture_hsk( void );
 uint8_t receive_byte( uint8_t *inout);
