+++ src/hexops.c
@@ -1,0 +1,3 @@
+#ifndef ARDUINO
+#include "hexops.cpp"
+#endif
