+++ src/printer.c
@@ -1,0 +1,3 @@
+#ifndef ARDUINO
+   #include "printer.cpp"
+#endif
