+++ src/printer.h
@@ -1,0 +1,19 @@
+/*
+ * printer.h
+ *
+ *  Created on: May 31, 2020
+ *      Author: brain
+ */
+
+#ifndef PRINTER_H
+#define PRINTER_H
+
+#include "config.h"
+#include "hexops.h"
+#include "registry.h"
+
+void prn_reset(void);
+void prn_register(void);
+void prn_init(void);
+
+#endif /* PRINTER_H */
