+++ src/rtc.c
@@ -1,0 +1,3 @@
+#ifndef ARDUINO
+   #include "rtc.cpp"
+#endif
