+++ src/serial.c
@@ -1,0 +1,3 @@
+#ifndef ARDUINO
+   #include "serial.cpp"
+#endif
